Fix focus issues on Windows 7 and 8

+++ code/Cairo Desktop/CairoDesktop.Interop/NativeMethods.cs
@@ -209,7 +209,7 @@ namespace CairoDesktop.Interop
             public int uCallbackMessage;
             public int uEdge;
             public RECT rc;
-            public IntPtr lParam;
+            public long lParam;
         }
 
         public enum ABMsg : int
